split VirtualService destinations analyzer
