+++ galley/pkg/config/analysis/analyzers/testdata/virtualservice_destinationhosts.yaml
@@ -1,0 +1,62 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: reviews
+  namespace: default
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: ServiceEntry
+metadata:
+  name: external-reviews
+  namespace: default
+spec:
+  hosts:
+  - external-reviews.org
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: reviews
+  namespace: default
+spec:
+  http:
+  - route:
+    - destination:  # This virtualservice has no validation errors (base case)
+        host: reviews
+        subset: v1
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: reviews-bogushost
+  namespace: default
+spec:
+  http:
+  - route:
+    - destination:
+        host: reviews-bogus # This host does not exist, should result in a validation error
+        subset: v1
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: reviews-fqdn
+  namespace: default
+spec:
+  http:
+  - route:
+    - destination:
+        host: reviews.default.svc.cluster.local # FQDN representation is valid and should not generate an error
+        subset: v1
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: reviews-external
+  namespace: default
+spec:
+  http:
+  - route:
+    - destination:
+        host: external-reviews.org  # Referring to a ServiceEntry host is valid and should not generate an error
+                                    # Since this is an "external" service, subset is omitted
